Refactored unit tests
